requirements.txt
customtkinter
PyYAML
deepdiff
